requirements.txt
# Core trading and data processing
ccxt==4.2.47
alpaca-py==0.24.0
websocket-client==1.7.0
python-binance==1.0.19

# Machine Learning and Analytics
scikit-learn==1.4.2
tensorflow==2.19.0
torch==2.3.0
transformers==4.39.3
numpy==1.26.4
pandas==2.2.2
finta==1.3
yfinance==0.2.28

# Database and Infrastructure
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
redis==5.0.4
celery==5.3.6

# API and Web Services
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
aiohttp==3.9.5
requests==2.31.0

# Configuration and Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
schedule==1.2.1
python-telegram-bot==21.1.1
loguru==0.7.2

# Monitoring and Performance
prometheus-client==0.20.0
grafana-client==4.0.0
APScheduler==3.10.4

# Data Science and Visualization
plotly==5.20.0
matplotlib==3.8.4
seaborn==0.13.2
jupyter==1.0.0
notebook==7.1.3

# Testing and Development
pytest==8.1.1
pytest-asyncio==0.23.6
black==24.4.0
flake8==7.0.0
mypy==1.10.0
